requirements.txt
dimod>=0.10.7
dwave-system>=1.10.0
pandas-datareader>=0.10.0
matplotlib>=3.3.4